requirements.txt
fastapi==0.136.0
httpx==0.28.1
jinja2==3.1.6
numpy==2.4.4
open-clip-torch==3.3.0
pandas==3.0.2
pillow==12.2.0
requests==2.32.5
scikit-learn==1.8.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.1
uvicorn[standard]==0.46.0
